requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
matplotlib>=3.8.0
pytorch-lightning>=2.2.0
tqdm>=4.66.0
joblib